[Refactor] replace zenlog by loguru due to 'warn' deprecation notice
